refactoring internal packages
