+++ internal/service/goutil/test/invalid/go.mod
