+++ internal/service/goutil/test/valid/go.mod
